{"schema":"decodex/commit/1","summary":"Extract frozen edit geometry owner","authority":"manual"}
